Added word2vec as an external dependency since the Google Code svn repository seems defunct, I point to the github clone of it. Additionally I changed the way we generate the corpus to create subcategories across items and runes for example. Finally updated processing of matches to support threading.
There does seem to be a bug that causes loss to go to nan and the only
thing I can pinpoint is the changes to process match, but even making
the training deterministic (except for thread order) makes the bug
randomly appear, so I'm not sure what the underlying problem is. I will
have to continue to try to pinpoint the issue.